+++ insert.d.ts
@@ -3,13 +3,18 @@ import type {ImmutableTree} from './types';
 export type InsertConflictHandler<T> =
   (existingTreeValue: T, value: T) => T;
 
-export function THROW(): never;
+export function onConflictThrowError(): never;
 
-export const NOOP:
+export const onConflictKeepTreeValue:
   <T>(treeValue: T, givenValue: T) => T;
 
-export const REPLACE:
+export const onConflictUseGivenValue:
   <T>(treeValue: T, givenValue: T) => T;
+
+// Aliases for backwards compatibility.
+export const NOOP: typeof onConflictKeepTreeValue;
+export const REPLACE: typeof onConflictUseGivenValue;
+export const THROW: typeof onConflictThrowError;
 
 export default function insert<T>(
   tree: ImmutableTree<T> | null,